Moved EventsForMotion resolver to motions file
