+++ src/data/resolvers/motions.ts
@@ -26,6 +26,52 @@ import {
 
 import { ProcessedEvent } from './colonyActions';
 
+const getMotionEvents = async (
+  votingReputationClient: ExtensionClient,
+  motionId: string,
+) => {
+  const motionStakedLogs = await getLogs(
+    votingReputationClient,
+    // @TODO Add missing types to colonyjs
+    // @ts-ignore
+    votingReputationClient.filters.MotionStaked(motionId, null, null, null),
+  );
+
+  const parsedMotionEvents = await Promise.all(
+    [...motionStakedLogs].map(async (log) => {
+      const parsedLog = votingReputationClient.interface.parseLog(log);
+      const { address, blockHash, blockNumber, transactionHash } = log;
+      const {
+        name,
+        values: { amount, ...rest },
+      } = parsedLog;
+      const stakeAmount = name === 'MotionStaked' ? amount : null;
+
+      return {
+        type: ActionsPageFeedType.NetworkEvent,
+        name,
+        values: {
+          ...rest,
+          stakeAmount,
+        },
+        createdAt: blockHash
+          ? await getBlockTime(votingReputationClient.provider, blockHash)
+          : 0,
+        emmitedBy: ClientType.VotingReputationClient,
+        address,
+        blockNumber,
+        transactionHash,
+      } as ProcessedEvent;
+    }),
+  );
+
+  const sortedMotionEvents = parsedMotionEvents.sort(
+    (firstEvent, secondEvent) => firstEvent.createdAt - secondEvent.createdAt,
+  );
+
+  return sortedMotionEvents;
+};
+
 export const motionsResolvers = ({
   colonyManager: { networkClient },
   colonyManager,
@@ -182,6 +228,19 @@ export const motionsResolvers = ({
         console.error('Could not fetch users vote reward');
         console.error(error);
         return null;
+      }
+    },
+    async eventsForMotion(_, { motionId, colonyAddress }) {
+      try {
+        const votingReputationClient = (await colonyManager.getClient(
+          ClientType.VotingReputationClient,
+          colonyAddress,
+        )) as ExtensionClient;
+
+        return await getMotionEvents(votingReputationClient, motionId);
+      } catch (error) {
+        console.error(error);
+        return [];
       }
     },
   },
